+++ API/run.py
@@ -1,0 +1,8 @@
+from __init__ import *
+from models import *
+
+
+
+if __name__ == '__main__':
+	create_tables()
+	app.run(debug=True)
